@@ -31,13 +31,13 @@
     "url": "https://github.com/flauc/ng-push/issues"
   },
   "devDependencies": {
-    "@angular/common": "5.0.0",
-    "@angular/compiler": "5.0.0",
-    "@angular/compiler-cli": "5.0.0",
-    "@angular/core": "5.0.0",
-    "@angular/http": "5.0.0",
-    "@angular/platform-browser": "5.0.0",
-    "@angular/platform-browser-dynamic": "5.0.0",
+    "@angular/common": "^6.0.1",
+    "@angular/compiler": "^6.0.1",
+    "@angular/compiler-cli": "^6.0.1",
+    "@angular/core": "^6.0.1",
+    "@angular/http": "^6.0.1",
+    "@angular/platform-browser": "^6.0.1",
+    "@angular/platform-browser-dynamic": "^6.0.1",
     "@compodoc/compodoc": "1.0.0-beta.10",
     "@types/jasmine": "2.5.38",
     "@types/node": "6.0.60",
@@ -57,7 +57,7 @@
     "karma-jasmine": "1.1.0",
     "karma-jasmine-html-reporter": "0.2.2",
     "lite-server": "2.3.0",
-    "node-sass": "4.5.2",
+    "node-sass": "4.9.0",
     "node-sass-tilde-importer": "1.0.0",
     "node-watch": "0.5.2",
     "protractor": "5.1.0",
@@ -66,9 +66,9 @@
     "systemjs": "0.20.12",
     "ts-node": "2.0.0",
     "tslint": "4.5.0",
-    "typescript": "2.4.2",
-    "rxjs": "5.5.2",
-    "zone.js": "0.8.14"
+    "typescript": ">=2.7.2 <2.8.0",
+    "rxjs": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "engines": {
     "node": ">=6.0.0"
